Refactor modularise algorithm
Refactor modularise algorithm
